requirements.txt
numpy>=1.16.1
easydict==1.9
opencv-python>=4.5.1.48
torch==2.2.2+cu121
torchvision==0.17.2+cu121
matplotlib~=3.7.0
pandas~=1.5.3
tqdm~=4.64.1
pillow~=9.4.0
pytorch_ssim